wasm_interp: move load and store tests to test_mem
